+++ lib/public/views/lhcPeriods/LhcPeriodsModel.js
@@ -12,7 +12,7 @@
  */
 
 import { Observable } from '/js/src/index.js';
-import { LhcPeriodsOverviewModel } from './Overview/LhcPeriodsModel.js';
+import { LhcPeriodsOverviewModel } from './Overview/LhcPeriodsOverviewModel.js';
 
 /**
  * LHC Periods model
@@ -29,7 +29,7 @@ export class LhcPeriodsModel extends Observable {
     }
 
     /**
-     * Retrieve every relevant LHC Period from the API
+     * Load LHC Periods overview data
      *
      * @returns {void}
      */
