+++ lib/public/views/lhcPeriods/Overview/LhcPeriodsOverviewModel.js
